refactors: client as module with basic statsd client and datadog client
