+++ aiostatsd/client/__init__.py
@@ -1,0 +1,2 @@
+from .basic import StatsdClient, StatsdClientBase  # noqa: F401
+from .datadog import DatadogClient  # noqa: F401
