+++ tests/test_client.py
@@ -9,7 +9,7 @@ from .testing import EventLoopMockMixin, RandomMockMixin, TimenowMockMixin
 class TestStatsdClientBase(IsolatedAsyncioTestCase, EventLoopMockMixin, RandomMockMixin):
     def setUp(self):
         self._create_loop_mock()
-        self._create_random_mock("aiostatsd.client.random")
+        self._create_random_mock("aiostatsd.client.basic.random")
 
     async def asyncSetUp(self):
         self.client = StatsdClientBase("localhost", 8125)
@@ -31,8 +31,8 @@ class TestStatsdClientBase(IsolatedAsyncioTestCase, EventLoopMockMixin, RandomMo
 class TestStatsdClient(IsolatedAsyncioTestCase, EventLoopMockMixin, RandomMockMixin, TimenowMockMixin):
     def setUp(self):
         self._create_loop_mock()
-        self._create_random_mock("aiostatsd.client.random")
-        self._create_timenow_mock("aiostatsd.client.time_now")
+        self._create_random_mock("aiostatsd.client.basic.random")
+        self._create_timenow_mock("aiostatsd.client.basic.time_now")
 
     async def asyncSetUp(self):
         self.client = StatsdClient("localhost", 8125)
